SmartGit index: don't link SmartGit to website
One likely way of the user to access the documentation is to reach it
from syntevo.com/smartgit/. If every word "SmartGit" is a link back the
SmartGit website, it hampers the flow of reading. IMHO it is sufficient
to have the SmartGit icon in the top bar being a link.